fixing minor errors

+++ db/migrate/20190708144647_create_multiple_honest_broker_types.rb
@@ -1,5 +1,5 @@
 class CreateMultipleHonestBrokerTypes < ActiveRecord::Migration[5.1]
   def change
-    change_column :users, :honest_broker, :bigint, default: 0
+    change_column :users, :honest_broker, :integer, default: 0
   end
 end
